refactor(db): abstract LISH read queries and row mappers into own module
